Add a utility that will provide an easy way to determine if a rebuild of a phar archive is necessary.
This is best suited for use with automated processes, to prevent the unnecessary distribution of archives that are rebuilt if a change is made, but the source files aren't affected.
When phar archives are rebuilt, it seems that their file contents change slightly and will register as an altered file, even though the source files held within haven't changed (e.g. doc files being updated in a package, or another library changes...latest git commit id changes, but the source files are the same).
This can save a decent amount of bandwidth, processing, and time - if used correctly this will ensure that phar archive rebuilds do not occur until it is actually necessary.